Neovim: switch to using LazyVim
**Why:**

I decided to try out LazyVim and it turns out it’s a really nice way of
managing my Neovim config. It comes with a lot of batteries included, so
I was able to remove a bunch of custom configuration for stuff like
Treesitter, LSP, Telescope, etc.

**What:**

Switch to using LazyVim based on the LazyVim starter[^1].

[^1]: https://github.com/LazyVim/starter

New setup is centred around these files, which handle various aspects of
config and plugins:

```
config
├── autocmds.lua	# Pretty
├── keymaps.lua 	# self-explanatory
├── lazy.lua    	# Initialise `lazy`
└── options.lua 	# `vim.g` & `vim.opt` stuff and abbreviations
plugins
├── add.lua			# Install, configure plugins not included in LazyVim
├── configure.lua	# Configure plugins that are included in LazyVim
└── disable.lua		# Disable plugins that are included in LazyVim
```

Additional changes:

* Add a stylua config for `.config/nvim`.
* Switch from using `markdownlint` to `markdownlint-cli2`.